+++ api/scpca_portal/test/test_validators.py
@@ -2,8 +2,9 @@ from django.test import TestCase
 
 from pydantic import ValidationError
 
-from scpca_portal.enums import Modalities
-from scpca_portal.validators import DatasetDataModel, ProjectDataModel
+from scpca_portal.enums import DatasetFormats, Modalities
+from scpca_portal.test.factories import ProjectFactory, SampleFactory
+from scpca_portal.validators import DatasetDataModel, DatasetDataResourceExistence, ProjectDataModel
 
 
 class TestProjectDataModel(TestCase):
@@ -82,3 +83,53 @@ class TestDatasetDataModel(TestCase):
             DatasetDataModel.model_validate(data)
 
         self.assertIn("Invalid sample ID format", str(context.exception))
+
+
+class TestDatasetDataResourceExistence(TestCase):
+    def test_validate(self):
+        project = ProjectFactory(scpca_id="SCPCP000001")
+        SampleFactory(scpca_id="SCPCS000001", project=project, has_single_cell_data=True)
+        SampleFactory(scpca_id="SCPCS000002", project=project, has_single_cell_data=True)
+        SampleFactory(scpca_id="SCPCS000003", project=project, has_spatial_data=True)
+
+        # no exceptions thrown
+        data = {
+            "SCPCP000001": {
+                "includes_bulk": True,
+                Modalities.SINGLE_CELL.value: ["SCPCS000001", "SCPCS000002"],
+                Modalities.SPATIAL.value: ["SCPCS000003"],
+            },
+        }
+        format = DatasetFormats.SINGLE_CELL_EXPERIMENT
+
+        DatasetDataResourceExistence.validate(data, format)  # no exception should be thrown here
+
+        # assert spatial samples cannot be requested with anndata format
+        data = {
+            "SCPCP000001": {
+                "includes_bulk": True,
+                Modalities.SINGLE_CELL.value: ["SCPCS000001", "SCPCS000002"],
+                Modalities.SPATIAL.value: ["SCPCS000003"],
+            },
+        }
+        format = DatasetFormats.ANN_DATA
+
+        with self.assertRaises(Exception) as e:
+            DatasetDataResourceExistence.validate(data, format)
+            self.assertEqual(str(e.exception), "No Spatial data for ANNDATA.")
+
+        # assert project id doesn't exist
+        data = {
+            "SCPCP999999": {
+                "includes_bulk": True,
+                Modalities.SINGLE_CELL.value: ["SCPCS000001", "SCPCS000002"],
+                Modalities.SPATIAL.value: ["SCPCS000003"],
+            },
+        }
+        format = DatasetFormats.SINGLE_CELL_EXPERIMENT
+
+        with self.assertRaises(Exception) as e:
+            DatasetDataResourceExistence.validate(data, format)
+            self.assertEqual(
+                str(e.exception), "The following projects do not exist: ['SCPCP999999']"
+            )
